create bud_calc, cell str utilitys
