+++ src/a10_bud_calc/_utils/str_a10.py
@@ -1,0 +1,10 @@
+def jmetrics_str() -> str:
+    return "jmetrics"
+
+
+def fund_take_str() -> str:
+    return "fund_take"
+
+
+def fund_give_str() -> str:
+    return "fund_give"
