+++ src/a10_bud_calc/test_/test_bud_calc_config.py
@@ -55,10 +55,12 @@ from src.a08_bud_atom_logic._utils.str_a08 import (
     sqlite_datatype_str,
 )
 from src.a08_bud_atom_logic.atom_config import get_atom_config_dict
-from src.a10_bud_calc.bud_calc_config import (
+from src.a10_bud_calc._utils.str_a10 import (
     jmetrics_str,
     fund_take_str,
     fund_give_str,
+)
+from src.a10_bud_calc.bud_calc_config import (
     get_bud_calc_config_filename,
     config_file_path,
     get_bud_calc_config_dict,
