+++ src/a11_deal_cell_logic/_utils/str_a10.py
@@ -1,0 +1,30 @@
+def ancestors_str() -> str:
+    return "ancestors"
+
+
+def celldepth_str() -> str:
+    return "celldepth"
+
+
+def deal_owner_name_str() -> str:
+    return "deal_owner_name"
+
+
+def mandate_str() -> str:
+    return "mandate"
+
+
+def budadjust_str() -> str:
+    return "budadjust"
+
+
+def budevent_facts_str() -> str:
+    return "budevent_facts"
+
+
+def found_facts_str() -> str:
+    return "found_facts"
+
+
+def boss_facts_str() -> str:
+    return "boss_facts"
