+++ src/a11_deal_cell_logic/cell.py
@@ -30,38 +30,6 @@ from copy import deepcopy as copy_deepcopy
 CELLNODE_QUOTA_DEFAULT = 1000
 
 
-def ancestors_str() -> str:
-    return "ancestors"
-
-
-def celldepth_str() -> str:
-    return "celldepth"
-
-
-def deal_owner_name_str() -> str:
-    return "deal_owner_name"
-
-
-def mandate_str() -> str:
-    return "mandate"
-
-
-def budadjust_str() -> str:
-    return "budadjust"
-
-
-def budevent_facts_str() -> str:
-    return "budevent_facts"
-
-
-def found_facts_str() -> str:
-    return "found_facts"
-
-
-def boss_facts_str() -> str:
-    return "boss_facts"
-
-
 @dataclass
 class CellUnit:
     ancestors: list[OwnerName] = None
