+++ src/a11_deal_cell_logic/test_cell/test_cell_.py
@@ -1,7 +1,7 @@
 from src.a01_word_logic.road import create_road
 from src.a04_reason_logic.reason_item import factunit_shop
 from src.a06_bud_logic.bud import budunit_shop
-from src.a11_deal_cell_logic.cell import (
+from src.a11_deal_cell_logic._utils.str_a10 import (
     ancestors_str,
     celldepth_str,
     deal_owner_name_str,
@@ -10,6 +10,8 @@ from src.a11_deal_cell_logic.cell import (
     budevent_facts_str,
     found_facts_str,
     boss_facts_str,
+)
+from src.a11_deal_cell_logic.cell import (
     CellUnit,
     cellunit_shop,
     CELLNODE_QUOTA_DEFAULT,
