+++ src/a11_deal_cell_logic/test_cell/test_cell_json.py
@@ -1,9 +1,7 @@
 from src.a02_finance_logic._utils.strs_a02 import quota_str
 from src.a06_bud_logic.bud import budunit_shop
 from src.a06_bud_logic._utils.str_a06 import event_int_str, penny_str
-from src.a11_deal_cell_logic.cell import (
-    cellunit_shop,
-    cellunit_get_from_dict,
+from src.a11_deal_cell_logic._utils.str_a10 import (
     ancestors_str,
     celldepth_str,
     deal_owner_name_str,
@@ -13,6 +11,7 @@ from src.a11_deal_cell_logic.cell import (
     found_facts_str,
     boss_facts_str,
 )
+from src.a11_deal_cell_logic.cell import cellunit_shop, cellunit_get_from_dict
 from src.a13_bud_listen_logic._utils.example_listen import (
     example_casa_clean_factunit as clean_factunit,
     example_casa_dirty_factunit as dirty_factunit,
