+++ src/a15_fisc_logic/test_fisc_tool/test_cell_tree__create__root.py
@@ -2,7 +2,11 @@ from src.a00_data_toolbox.file_toolbox import save_file, open_json, count_dirs_f
 from src.a02_finance_logic._utils.strs_a02 import celldepth_str, quota_str
 from src.a02_finance_logic.deal import DEFAULT_CELLDEPTH
 from src.a06_bud_logic._utils.str_a06 import event_int_str, penny_str
-from src.a11_deal_cell_logic.cell import deal_owner_name_str, ancestors_str
+from src.a11_deal_cell_logic._utils.str_a10 import (
+    ancestors_str,
+    celldepth_str,
+    deal_owner_name_str,
+)
 from src.a12_hub_tools.hub_path import (
     create_fisc_json_path,
     create_fisc_owners_dir_path,
