+++ src/a15_fisc_logic/test_fisc_tool/test_cell_tree_budevents.py
@@ -1,5 +1,5 @@
 from src.a00_data_toolbox.file_toolbox import open_json
-from src.a11_deal_cell_logic.cell import budevent_facts_str
+from src.a11_deal_cell_logic._utils.str_a10 import budevent_facts_str
 from src.a12_hub_tools.hub_path import create_cell_json_path, create_budevent_path
 from src.a12_hub_tools.hub_tool import save_arbitrary_budevent, cellunit_add_json_file
 from src.a15_fisc_logic.fisc_tool import load_cells_budevent
