+++ src/a19_world_logic/test_08_create_balances/test_world_etl08_00_create_root_deal_nodes.py
@@ -7,7 +7,7 @@ from src.a00_data_toolbox.file_toolbox import (
 from src.a02_finance_logic._utils.strs_a02 import celldepth_str, quota_str
 from src.a02_finance_logic.deal import DEFAULT_CELLDEPTH
 from src.a06_bud_logic._utils.str_a06 import event_int_str, penny_str
-from src.a11_deal_cell_logic.cell import deal_owner_name_str, ancestors_str
+from src.a11_deal_cell_logic._utils.str_a10 import ancestors_str, deal_owner_name_str
 from src.a12_hub_tools.hub_path import (
     create_fisc_json_path,
     create_fisc_owners_dir_path,
